+++ src/main/java/nextstep/subway/exception/ControllerExceptionHandler.java
@@ -1,0 +1,27 @@
+package nextstep.subway.exception;
+
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import javax.persistence.EntityNotFoundException;
+import java.util.NoSuchElementException;
+
+@ControllerAdvice
+public class ControllerExceptionHandler {
+    @ExceptionHandler({EntityNotFoundException.class, NoSuchElementException.class})
+    public ResponseEntity handleNotFoundException() {
+        return ResponseEntity.noContent().build();
+    }
+
+    @ExceptionHandler({IllegalArgumentException.class})
+    public ResponseEntity handleIllegalArgsException() {
+        return ResponseEntity.badRequest().build();
+    }
+
+    @ExceptionHandler(DataIntegrityViolationException.class)
+    public ResponseEntity handleDataIntegrityViolationException() {
+        return ResponseEntity.badRequest().build();
+    }
+}
